chore(unionlabs): move types to unionlabs
