+++ lib/unionlabs/src/aptos/bit_vec.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Default, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+pub struct BitVec {
+    #[serde(with = "serde_bytes")]
+    inner: Vec<u8>,
+}
