+++ light-clients/movement/ics08-movement/Cargo.toml
@@ -15,8 +15,11 @@ bcs = { git = "https://github.com/aptos-labs/bcs.git", rev = "d31fab9d81748e2594
 sha2 = { workspace = true, features = [ "std" ] }
 sha3 = { workspace = true }
 serde = { workspace = true, features = [ "derive" ] }
-serde_bytes = "0.11.6"
 thiserror = { workspace = true }
+ics008-wasm-client     = { workspace = true }
+unionlabs = { workspace = true }
+cosmwasm-std           = { workspace = true, features = ["abort"] }
+protos = { workspace = true }
 
 [lints]
 workspace = true
